finished app

+++ the_bank/api/routes.py
@@ -1,6 +1,7 @@
 """Import libraries."""
 from flask import Blueprint, request, jsonify
-from the_bank import db
+from flask_praetorian import auth_required
+from the_bank import db, guard
 from the_bank.models import Account
 
 api = Blueprint("api", __name__)
@@ -21,7 +22,9 @@ def open_account():
     account = Account.query.filter_by(holder=holder).first()
     if account:
         return jsonify({"error": "Account already exists"})
-    account = Account(holder=holder)
+    password = request.json.get("password")
+    hashed_password = guard.hash_password(password)
+    account = Account(holder=holder, password=hashed_password)
     db.session.add(account)
     db.session.commit()
     return (
@@ -32,32 +35,48 @@ def open_account():
     )
 
 
-@api.route("/account/<holder>", methods=["GET"])
-def view_account(holder):
+@api.route("/account", methods=["PUT"])
+def view_account():
     """View an account."""
+    holder = request.json.get("holder")
     account = Account.query.filter_by(holder=holder).first()
     if account:
+        password = request.json.get("password")
+        account = guard.authenticate(holder, password)
+        token = guard.encode_jwt_token(account)
         return (
-            jsonify({"holder": account.holder, "balance": account.balance}),
+            jsonify(
+                {
+                    "token": token,
+                    "holder": account.holder,
+                    "balance": account.balance,
+                }
+            ),
             200,
         )
     return jsonify({"error": "Account does not exist"})
 
 
-@api.route("/account/<holder>", methods=["DELETE"])
-def close_account(holder):
+@api.route("/account", methods=["DELETE"])
+def close_account():
     """Close an account."""
+    holder = request.json.get("holder")
+    print(f"\n{holder}\n")
     account = Account.query.filter_by(holder=holder).first()
     if not account:
         return jsonify({"error": "Account does not exist"})
     if account.balance > 0:
         return jsonify({"error": "Remove balance before closing account"})
-    db.session.delete(account)
-    db.session.commit()
-    return jsonify({"message": "The account has been closed"})
+    password = request.json.get("password")
+    account = guard.authenticate(holder, password)
+    if account:
+        db.session.delete(account)
+        db.session.commit()
+        return jsonify({"message": "The account has been closed"})
 
 
 @api.route("/account/<holder>/deposit", methods=["POST"])
+@auth_required
 def deposit(holder):
     """Deposit to account of a given holder."""
     account = Account.query.filter_by(holder=holder).first()
@@ -76,6 +95,7 @@ def deposit(holder):
 
 
 @api.route("/account/<holder>/withdraw", methods=["POST"])
+@auth_required
 def withdraw(holder):
     """Withdraw from account of a given holder."""
     account = Account.query.filter_by(holder=holder).first()
